backend/requirements.txt
pywin32==302; sys_platform == 'win32'
aiofiles
aiohttp
aiosignal
altair
annotated-types
anyio
async-timeout
attrs
Automat
beautifulsoup4
blinker
blis
Brotli
bs4
cachetools
camelot-py
catalogue
certifi
cffi
chardet
charset-normalizer
ci-info
click
colorama
confection
configobj
configparser
constantly
cryptography
cssselect
cssselect2
cymem
dataclasses-json
diskcache
distro
docx2pdf
docx2txt
et-xmlfile
etelemetry
fake-headers
filelock
fitz
Flask
Flask-Cors
fonttools
frontend
frozenlist
fsspec
future
fuzzywuzzy
ghostscript
gitdb
GitPython
gobject
googlesearch-python
greenlet
h11
html-table-extractor
html2text
html5lib
httplib2
huggingface-hub
hyperlink
idna
importlib-metadata
incremental
inscriptis
isodate
itemadapter
itemloaders
itsdangerous
Jinja2
jmespath
joblib
JPype1
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
langchain
langcodes
langsmith
llama_cpp_python
looseversion
lxml
markdown-it-py
MarkupSafe
marshmallow
mdurl
mpmath
multidict
murmurhash
mypy-extensions
networkx
nipype
nltk
numpy
opencv-python
openpyxl
packaging
pandas
parsel
pathlib
pathy
pdf2image
pdfkit
pdfminer.six
pdfplumber
pdfquery
Pillow
pip
pptxer
preshed
prettytable
Protego
protobuf
prov
pyarrow
pyasn1
pyasn1-modules
pycparser
pydantic
pydantic_core
pydeck
PyDispatcher
pydot
pydyf
Pygments
Pympler
PyMuPDF
PyMuPDFb
pyOpenSSL
pyparsing
pypdf
PyPDF2
pypdfium2
pyphen
pyquery
pytesseract
python-dateutil
python-docx
python-pptx
pytz
pytz-deprecation-shim
pyxlsb
pyxnat
PyYAML
queuelib
rdflib
referencing
regex
reportlab
requests
requests-file
rich
roman
rpds-py
safetensors
scikit-learn
scipy
Scrapy
sentencepiece
service-identity
setuptools
simplejson
six
smart-open
smmap
sniffio
soupsieve
spacy
spacy-legacy
spacy-loggers
SQLAlchemy
srsly
starlette
streamlit
sympy
tabula
tabula-py
tabulate
tenacity
texttable
thinc
threadpoolctl
tika
tinycss2
tldextract
tokenizers
toml
toolz
torch
tornado
tqdm
traits
transformers
typer
typing_extensions
typing-inspect
tzdata
tzlocal
urllib3
uvicorn
validators
w3lib
wasabi
watchdog
wcwidth
weasyprint
webencodings
Werkzeug
wheel
wkhtmltopdf
wordninja
xlrd
XlsxWriter
xlwings
yarl
zipp
